Break up interchange::avro into constituent modules
This breaks up `avro.rs` (which already contained one inline module) into smaller modules that are
each basically responsible for only one thing. There are no code changes, only copy/paste.

"Basically" because our envelope logic currently depends on both encoding and decoding logic. There
are definitely useful forward steps from this work to make the dependencies between the various
parts stand on their own a bit more, but at least now there are explicit `use *` lines to document
the interdependence.

This change came about because I was trying to understand exactly how flexible our decode logic is
and what would go into refactoring it to support injecting it into our upsert operator, and the
functionality for all these basically unrelated thigns were interwoven.
